feat: move tree item types to types file
