+++ src/connect/components/index.ts
@@ -1,11 +1,14 @@
 export * from './breadcrumbs';
 export * from './divider';
+export * from './drive-settings-form';
+export * from './drive-settings-modal';
 export * from './drive-settings-select';
 export * from './drive-view';
 export * from './dropdown-menu';
 export * from './modal';
 export * from './search-bar';
 export * from './sidebar';
+export * from './status-indicator';
 export * from './toggle';
 export * from './tree-view';
 export * from './tree-view-item';
