+++ src/connect/types/ui.ts
@@ -1,3 +1,5 @@
+import { ConnectDropdownMenuItem } from '@/connect';
+
 export type TreeItemType =
     | 'folder'
     | 'file'
@@ -16,3 +18,29 @@ export type TreeItemStatus =
     | 'available-offline'
     | 'syncing'
     | 'offline';
+
+export type SyncStatus = 'not-synced-yet' | 'syncing' | 'synced';
+
+export type SharingType = 'private' | 'shared' | 'public';
+
+export type BaseTreeItem = {
+    id: string;
+    path: string;
+    label: string;
+    type: TreeItemType;
+    error?: Error;
+    status?: TreeItemStatus;
+    isConnected?: boolean;
+    options?: ConnectDropdownMenuItem[];
+    syncStatus?: SyncStatus;
+    sharingType?: SharingType;
+    expanded?: boolean;
+};
+
+export type UITreeItem = {
+    action?: TreeItemAction;
+    expanded?: boolean;
+    isSelected?: boolean;
+};
+
+export type TreeItem = BaseTreeItem & UITreeItem;
