+++ src/connect/utils/index.ts
@@ -1,2 +1,3 @@
+export * from './mocks';
 export * from './path';
 export * from './tree-item';
